+++ app/models/person.rb
@@ -56,7 +56,7 @@ class Person < ApplicationRecord
   accepts_nested_attributes_for :family_relationships, reject_if: -> (p) { p[:related_contact_id].blank? }, allow_destroy: true
   accepts_nested_attributes_for :facebook_accounts, reject_if: -> (p) { p[:username].blank? }, allow_destroy: true
   accepts_nested_attributes_for :twitter_accounts, reject_if: -> (p) { p[:screen_name].blank? }, allow_destroy: true
-  accepts_nested_attributes_for :linkedin_accounts, reject_if: -> (p) { p[:url].blank? }, allow_destroy: true
+  accepts_nested_attributes_for :linkedin_accounts, reject_if: -> (p) { p[:public_url].blank? }, allow_destroy: true
   accepts_nested_attributes_for :pictures, reject_if: -> (p) { p[:image].blank? && p[:image_cache].blank? }, allow_destroy: true
   accepts_nested_attributes_for :websites, reject_if: -> (p) { p[:url].blank? }, allow_destroy: true
 
@@ -109,7 +109,8 @@ class Person < ApplicationRecord
       linkedin_accounts_attributes: [
         :_destroy,
         :id,
-        :url
+        :public_url,
+        :uuid
       ],
       phone_number: :number,
       phone_numbers_attributes: [