+++ app/serializers/person/linkedin_account_serializer.rb
@@ -1,11 +1,7 @@
 class Person::LinkedinAccountSerializer < ApplicationSerializer
   type :linkedin_accounts
 
-  attributes :authenticated,
-             :created_at,
-             :first_name,
-             :last_name,
+  attributes :created_at,
              :public_url,
-             :remote_id,
              :updated_at
 end